Compose - eliminated most of errors with user manipulation.

+++ app/src/main/java/com/simenko/qmapp/ui/user/repository/UserManager.kt
@@ -10,6 +10,8 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
 import javax.inject.Singleton
+import kotlin.coroutines.resume
+import kotlin.coroutines.suspendCoroutine
 
 private const val REGISTERED_USER = "registered_user"
 private const val PASSWORD_SUFFIX = "password"
@@ -36,13 +38,40 @@ class UserManager @Inject constructor(
         get() = storage.getString(REGISTERED_USER)
 
     fun isUserLoggedIn(): Boolean {
-        Log.d(TAG, "isUserLoggedIn: ${storage.getString(REGISTERED_USER)}")
-        return userDataRepository.username != null
+        val result = userDataRepository.username != null
+        if (result) _userState.value = Event(UserLoggedInState)
+        return result
     }
 
-    fun isUserRegistered(): Boolean {
-        Log.d(TAG, "isUserRegistered: ${storage.getString(REGISTERED_USER)}")
-        return storage.getString(REGISTERED_USER).isNotEmpty()
+    suspend fun isUserRegistered() = suspendCoroutine { continuation ->
+        val registeredUser = storage.getString(REGISTERED_USER)
+        val registeredPassword = storage.getString("$username$PASSWORD_SUFFIX")
+
+        if (registeredUser.isEmpty() || registeredPassword.isEmpty()) {
+            _userState.value = Event(UserInitialState)
+            continuation.resume(false)
+        } else {
+            auth.signInWithEmailAndPassword(registeredUser, registeredPassword)
+                .addOnCompleteListener { task ->
+                    if (task.isSuccessful) {
+                        auth.signOut()
+                        _userState.value = Event(UserRegisteredState("user is registered - verified with Firebase"))
+                        continuation.resume(true)
+                    } else {
+                        when (task.exception) {
+                            is FirebaseNetworkException -> {
+                                _userState.value = Event(UserRegisteredState("No network - just continue with Storage"))
+                                continuation.resume(true)
+                            }
+
+                            else -> {
+                                _userState.value = Event(UserInitialState)
+                                continuation.resume(false)
+                            }
+                        }
+                    }
+                }
+        }
     }
 
     fun registerUser(username: String, password: String) {
@@ -53,51 +82,54 @@ class UserManager @Inject constructor(
                     storage.setString("$username$PASSWORD_SUFFIX", password)
                     userJustLoggedIn(storage.getString(REGISTERED_USER))
                     _userState.value = Event(UserLoggedInState)
-                    Log.d(TAG, "createUserWithEmail:success")
                 } else {
-                    when(task.exception) {
+                    when (task.exception) {
                         is FirebaseAuthUserCollisionException -> {
+                            storage.setString(REGISTERED_USER, username)
+                            storage.setString("$username$PASSWORD_SUFFIX", password)
                             _userState.value = Event(UserRegisteredState(task.exception?.message))
                         }
+
                         else -> {
                             _userState.value = Event(UserErrorState(task.exception?.message))
                         }
                     }
-                    Log.w(TAG, "createUserWithEmail:failure", task.exception)
                 }
             }
     }
 
     fun loginUser(username: String, password: String) {
-        val registeredUser = this.username
-        val registeredPassword = storage.getString("$username$PASSWORD_SUFFIX")
+        if (username.isNotEmpty() && password.isNotEmpty())
+            auth.signInWithEmailAndPassword(username, password)
+                .addOnCompleteListener { task ->
+                    if (task.isSuccessful) {
+                        Log.d(TAG, "loginUser: ${auth.currentUser?.email}")
+                        storage.setString(REGISTERED_USER, auth.currentUser?.email ?: "no mail")
+                        storage.setString("$username$PASSWORD_SUFFIX", password)
+                        userJustLoggedIn(username)
+                        _userState.value = Event(UserLoggedInState)
+                    } else {
+                        when (task.exception) {
+                            is FirebaseNetworkException -> {
+                                if (this.username == username && storage.getString("$username$PASSWORD_SUFFIX") == password) {
+                                    userJustLoggedIn(username)
+                                    _userState.value = Event(UserLoggedInState)
+                                } else if (this.username != username) {
+                                    _userState.value = Event(UserErrorState("Wrong email"))
+                                } else if (storage.getString("$username$PASSWORD_SUFFIX") != password) {
+                                    _userState.value = Event(UserErrorState("Wrong password"))
+                                }
+                            }
 
-        auth.signInWithEmailAndPassword(storage.getString(REGISTERED_USER), storage.getString("$username$PASSWORD_SUFFIX"))
-            .addOnCompleteListener { task ->
-                if (task.isSuccessful) {
-                    storage.setString(REGISTERED_USER, auth.currentUser?.email ?: "no mail")
-                    storage.setString("$username$PASSWORD_SUFFIX", password)
-                    _userState.value = Event(UserLoggedInState)
-                    userJustLoggedIn(username)
-                } else {
-                    when (task.exception) {
-                        is FirebaseNetworkException -> {
-                            if (this.username == username && storage.getString("$username$PASSWORD_SUFFIX") == password) {
-                                _userState.value = Event(UserLoggedInState)
-                                userJustLoggedIn(username)
-                            } else if (this.username != username) {
-                                _userState.value = Event(UserErrorState("Wrong email"))
-                            } else if (storage.getString("$username$PASSWORD_SUFFIX") != password) {
-                                _userState.value = Event(UserErrorState("Wrong password"))
+                            else -> {
+                                _userState.value = Event(UserErrorState(task.exception?.message))
                             }
                         }
-                        else -> {
-                            _userState.value = Event(UserErrorState(task.exception?.message))
-                        }
                     }
-                    Log.d(TAG, "loginUser/error: ${task.exception}")
                 }
-            }
+        else {
+            _userState.value = Event(UserErrorState("Email and Password fields should not be empty"))
+        }
     }
 
     fun logout() {
@@ -113,26 +145,29 @@ class UserManager @Inject constructor(
         }
     }
 
-    fun unregister() {
-        auth.signInWithEmailAndPassword(storage.getString(REGISTERED_USER), storage.getString("$username$PASSWORD_SUFFIX"))
-            .addOnCompleteListener { task1 ->
-                if (task1.isSuccessful) {
-                    auth.currentUser?.delete()?.addOnCompleteListener { task2 ->
-                        if (task2.isSuccessful) {
-                            val username = storage.getString(REGISTERED_USER)
-                            storage.setString(REGISTERED_USER, "")
-                            storage.setString("$username$PASSWORD_SUFFIX", "")
-                            logout()
-                            _userState.value = Event(UserInitialState)
-                            Log.d(TAG, "deleteUser:success")
-                        } else {
-                            _userState.value = Event(UserErrorState(task2.exception?.message ?: "Unknown error"))
+    fun deleteProfile(username: String, password: String) {
+        if (username.isNotEmpty() && password.isNotEmpty())
+            auth.signInWithEmailAndPassword(username, password)
+                .addOnCompleteListener { task1 ->
+                    if (task1.isSuccessful) {
+                        auth.currentUser?.delete()?.addOnCompleteListener { task2 ->
+                            if (task2.isSuccessful) {
+                                storage.setString(REGISTERED_USER, "")
+                                storage.setString("$username$PASSWORD_SUFFIX", "")
+                                logout()
+                                _userState.value = Event(UserInitialState)
+                                Log.d(TAG, "deleteUser:success")
+                            } else {
+                                _userState.value = Event(UserErrorState(task2.exception?.message ?: "Unknown error"))
+                            }
                         }
+                    } else {
+                        _userState.value = Event(UserErrorState(task1.exception?.message ?: "Unknown error"))
                     }
-                } else {
-                    _userState.value = Event(UserErrorState(task1.exception?.message ?: "Unknown error"))
                 }
-            }
+        else {
+            _userState.value = Event(UserErrorState("Email and Password fields should not be empty"))
+        }
     }
 
     private fun userJustLoggedIn(username: String) {
